convert liveblog-adverts.spec to typescript
